Freezing remote theme to specific commit
The most recent released version of jekyll/minima (2.5.1) does not
include support custom-head.html; that support is just sitting on
master. However, I don't want to just be swept along with whatever
future changes happen in master, so I'm trying to specify a specific
commit of the theme I want to use (for now, just the hash of HEAD at the
time of this commit). That way, I have a lever to pull to update when I
want to.